update tests for shiftCRUDContainer
