+++ desktop/src/components/forms/Shift/FullShift/fullShift.test.js
@@ -6,8 +6,7 @@ import { Formik } from 'formik';
 
 import {
   FullShift,
-  ANALYZE_SHIFT_FULL_SHIFT_RESET_BUTTON_ID,
-  
+  ANALYZE_SHIFT_FULL_SHIFT_RESET_BUTTON_ID,  
   ANALYZE_SHIFT_FULL_SHIFT_REMOVE_ACTIVITY_BUTTON_ID,
   ANALYZE_SHIFT_FULL_SHIFT_ADD_ACTIVITY_BUTTON_ID,
   ANALYZE_SHIFT_FULL_SHIFT_PROJECT_FIELD_ID
